+++ appSearch/build.gradle
@@ -3,7 +3,7 @@ plugins {
 }
 
 android {
-    namespace "com.louis993546.metro.app_search"
+    namespace "com.louis993546.metro.appSearch"
 }
 
 dependencies {
