+++ appSearch/proguard-rules.pro
