+++ appSearch/src/main/AndroidManifest.xml
