+++ build.gradle
@@ -35,7 +35,7 @@ apiValidation {
             "metro_settings",
             "settings",
             "appDrawer",
-            "app_search",
+            "appSearch",
             "calendar",
             "launcher",
             "apps",
