+++ settings.gradle
@@ -40,7 +40,7 @@ include ":browser"
 include ":radio"
 include ":calendar"
 include ":vertical_tiles_grid"
-include ":app_search"
+include ":appSearch"
 include ":appDrawer"
 include ":launcher"
 include ":apps"
